Factor out functions related to objects
